refactor: hook의 로직 대부분을 다시 컴포넌트 내부로 이동시킴
