+++ src/components/rangeSlider/RangeSlider.tsx
@@ -3,11 +3,20 @@ import RangeSliderBackground from './RangeSliderBackground';
 import RangeSliderHandle from './RangeSliderHandle';
 import { HandlesProps } from 'rc-slider/lib/Handles';
 import S from './RangeSlider.style';
+import { useMemo } from 'react';
+import RangeSliderMarkText from './RangeSliderMarkText';
 
-export type RangeSliderProps = Pick<
-  SliderProps,
-  'min' | 'max' | 'step' | 'defaultValue' | 'value' | 'onChange' | 'marks'
->;
+export type RangeValueType = [number, number];
+
+export type RangeSliderProps = Required<
+  Pick<SliderProps, 'min' | 'max' | 'step'>
+> & {
+  minMarkPrefix?: string;
+  maxMarkPostfix?: string;
+  markStep: number;
+  value: RangeValueType;
+  onChange: (value: RangeValueType) => void;
+};
 
 const handleRender: HandlesProps['handleRender'] = (origin, props) => (
   <S.CustomHandleContainer {...origin.props} {...props}>
@@ -19,11 +28,38 @@ const RangeSlider = ({
   min,
   max,
   step,
-  defaultValue,
   value,
-  marks,
+  minMarkPrefix,
+  maxMarkPostfix,
+  markStep,
   onChange,
 }: RangeSliderProps) => {
+  const marks = useMemo(() => {
+    const ret: SliderProps['marks'] = {};
+
+    for (let index = min; index <= max; index += markStep)
+      ret[index] = <RangeSliderMarkText label={`${index}`} />;
+
+    const minMarkIndex = Math.min(
+      ...Object.keys(ret).map(item => Number(item)),
+    );
+    const maxMarkIndex = Math.max(
+      ...Object.keys(ret).map(item => Number(item)),
+    );
+
+    if (minMarkPrefix)
+      ret[minMarkIndex] = (
+        <RangeSliderMarkText label={`${minMarkPrefix}${minMarkIndex}`} />
+      );
+
+    if (maxMarkPostfix)
+      ret[maxMarkIndex] = (
+        <RangeSliderMarkText label={`${maxMarkIndex}${maxMarkPostfix}`} />
+      );
+
+    return ret;
+  }, [min, max, markStep, minMarkPrefix, maxMarkPostfix]);
+
   return (
     <S.EntireContainer>
       <S.BackgroundContainer>
@@ -33,16 +69,13 @@ const RangeSlider = ({
         <Slider
           range
           pushable
-          {...{
-            min,
-            max,
-            step,
-            defaultValue,
-            value,
-            marks,
-            onChange,
-            handleRender,
-          }}
+          min={min}
+          max={max}
+          step={step}
+          value={value}
+          marks={marks}
+          handleRender={handleRender}
+          onChange={onChange as SliderProps['onChange']}
         />
       </S.SliderContainer>
     </S.EntireContainer>
